requirements.txt
# Core dependencies
langchain-core==0.3.62
langchain-google-genai==2.1.5
pydantic==2.11.4
pydantic_core==2.33.2

# Testing
pytest==8.3.5
pluggy==1.5.0

# Language and text processing
langdetect==1.0.9
babel==2.17.0

# HTTP and API
requests==2.32.3
httpx==0.28.1
httpcore==1.0.9
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# LangGraph related
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
typing_extensions==4.13.2
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
packaging==24.2
pyparsing==3.2.3
six==1.17.0
tqdm==4.67.1

# Protocol and serialization
protobuf==5.29.4
proto-plus==1.26.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.18
ormsgpack==1.9.1

# Type checking and validation
annotated-types==0.7.0
typing-inspection==0.4.0
mypy==1.9.0

# Optional dependencies (can be removed if not used)
anyio==4.9.0
cachetools==5.5.2
h11==0.16.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
uritemplate==4.1.1
xxhash==3.5.0
zstandard==0.23.0
